+++ Clothes Shop ERP/Core/Modules/Settings/Payment/UcPaymentMethods.cs
@@ -20,7 +20,6 @@ namespace Clothes_Shop_ERP.modlestore
             InitializeComponent();
             GetData();
             gridView1.OptionsView.ShowGroupPanel = false;
-            gridView1.OptionsCustomization.AllowSort = false;
             Sett.CenterColumns(gridView1);
             ApplyLanguage();
         }
@@ -56,7 +55,8 @@ namespace Clothes_Shop_ERP.modlestore
                 menu.Items.Add(LocalizationManager.T("Shared_MenuEdit"), null, (s, ev) => EditSelected());
                 menu.Items.Add(LocalizationManager.T("PaymentMethods_MenuToggleCash"), null, (s, ev) => ToggleCash());
                 menu.Items.Add(LocalizationManager.T("Shared_MenuActivateDeactivate"), null, (s, ev) => ToggleActive());
-                menu.Items.Add(LocalizationManager.T("Shared_MenuDelete"), null, (s, ev) => DeleteSelected());
+                if (PermissionManager.CanDelete("PaymentMethods"))
+                    menu.Items.Add(LocalizationManager.T("Shared_MenuDelete"), null, (s, ev) => DeleteSelected());
             }
 
             menu.Show(gridControl1, e.Location);
